# openapi/openapi_sert_subkey.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cruscotto Sert pagoPA backend service API (Subscription Key)",
    "version": "1.1.2-9-feature-sert-mass-search-api"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/sub/api/workflows/{nav}/{pa-emittente}": {
      "get": {
        "tags": [
          "Visualizzazione Dettagli"
        ],
        "operationId": "getWorkflows",
        "parameters": [
          {
            "name": "nav",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pa-emittente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "4XX": {
            "description": "Client error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowObjectDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sub/api/transfers/{nav}/{pa-emittente}/{token}": {
      "get": {
        "tags": [
          "Visualizzazione Dettagli"
        ],
        "operationId": "getTransfers",
        "parameters": [
          {
            "name": "nav",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pa-emittente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "4XX": {
            "description": "Client error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferPaymentDTO"
                }
              }
            }
          }
        }
      }
    },
    "/sub/api/token/{token}": {
      "get": {
        "tags": [
          "Visualizzazione Dettagli"
        ],
        "operationId": "getTokenInfo",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "4XX": {
            "description": "Client error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenInfoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/sub/api/search": {
      "get": {
        "tags": [
          "Ricerca delle posizioni debitorie"
        ],
        "operationId": "search",
        "parameters": [
          {
            "name": "pa",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nav",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iuv",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idCarrello",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "info",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "4XX": {
            "description": "Client error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionPaymentExtraDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sub/api/position/{nav}/{pa-emittente}": {
      "get": {
        "tags": [
          "Visualizzazione posizione debitoria"
        ],
        "operationId": "getPosition",
        "parameters": [
          {
            "name": "nav",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pa-emittente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "4XX": {
            "description": "Client error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionPaymentDTO"
                }
              }
            }
          }
        }
      }
    },
    "/sub/api/extra/{token}": {
      "get": {
        "tags": [
          "Visualizzazione Dettagli"
        ],
        "operationId": "getExtraInfo",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "4XX": {
            "description": "Client error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailWithCause"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraInfoResponseDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetailWithCause": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titolo del problema"
          },
          "status": {
            "type": "integer",
            "description": "Codice di stato HTTP",
            "format": "int32"
          },
          "detail": {
            "type": "string",
            "description": "Dettaglio del problema"
          }
        },
        "description": "Rappresentazione dettagliata di un problema con causa"
      },
      "WorkflowObjectDTO": {
        "type": "object",
        "properties": {
          "insertedtimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "tipoevento": {
            "type": "string"
          },
          "sottotipoevento": {
            "type": "string"
          },
          "outcome": {
            "type": "string"
          },
          "faultcode": {
            "type": "string"
          },
          "corr-id": {
            "type": "string"
          },
          "event-id": {
            "type": "string"
          }
        }
      },
      "PositionPaymentInfoDTO": {
        "type": "object",
        "properties": {
          "nav": {
            "type": "string"
          },
          "iuv": {
            "type": "string"
          },
          "pa-emittente": {
            "type": "string"
          },
          "creditor-reference-id": {
            "type": "string"
          },
          "last-event": {
            "type": "string",
            "format": "date-time"
          },
          "is-cached": {
            "type": "boolean"
          }
        }
      },
      "TransferObjectDTO": {
        "type": "object",
        "properties": {
          "idTransfer": {
            "type": "integer",
            "format": "int32"
          },
          "iban": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "type-transfer": {
            "type": "string"
          },
          "pa-fiscal-code": {
            "type": "string"
          }
        }
      },
      "TransferPaymentDTO": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "transfers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferObjectDTO"
            }
          },
          "position-info": {
            "$ref": "#/components/schemas/PositionPaymentInfoDTO"
          },
          "transfers-count": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ActorsDTO": {
        "type": "object",
        "properties": {
          "psp": {
            "type": "string"
          },
          "station": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "psp-description": {
            "type": "string"
          },
          "pt-pa": {
            "type": "string"
          },
          "pt-pa-description": {
            "type": "string"
          },
          "pt-psp": {
            "type": "string"
          },
          "pt-psp-description": {
            "type": "string"
          }
        }
      },
      "AmountDTO": {
        "type": "object",
        "properties": {
          "fee": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PayedDTO": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "payment-born": {
            "type": "string",
            "format": "date-time"
          },
          "payed-date": {
            "type": "string",
            "format": "date-time"
          },
          "multi-outcome": {
            "type": "boolean"
          }
        }
      },
      "PaymentInfoDTO": {
        "type": "object",
        "properties": {
          "touchpoint": {
            "type": "string"
          },
          "payment-method": {
            "type": "string"
          },
          "is-dw": {
            "type": "boolean"
          },
          "is-gpd": {
            "type": "boolean"
          },
          "is-standin": {
            "type": "boolean"
          },
          "is-cart": {
            "type": "boolean"
          }
        }
      },
      "TokenInfoDTO": {
        "type": "object",
        "properties": {
          "payed": {
            "$ref": "#/components/schemas/PayedDTO"
          },
          "actors": {
            "$ref": "#/components/schemas/ActorsDTO"
          },
          "amount": {
            "$ref": "#/components/schemas/AmountDTO"
          },
          "position-info": {
            "$ref": "#/components/schemas/PositionPaymentInfoDTO"
          },
          "is-payed-token": {
            "type": "boolean"
          },
          "payment-info": {
            "$ref": "#/components/schemas/PaymentInfoDTO"
          }
        }
      },
      "PositionPaymentExtraDTO": {
        "type": "object",
        "properties": {
          "nav": {
            "type": "string"
          },
          "match": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pa-emittente": {
            "type": "string"
          }
        }
      },
      "PositionPaymentDTO": {
        "type": "object",
        "properties": {
          "tokens": {
            "type": "integer",
            "format": "int32"
          },
          "payed": {
            "$ref": "#/components/schemas/PayedDTO"
          },
          "actors": {
            "$ref": "#/components/schemas/ActorsDTO"
          },
          "amount": {
            "$ref": "#/components/schemas/AmountDTO"
          },
          "position-info": {
            "$ref": "#/components/schemas/PositionPaymentInfoDTO"
          },
          "all-tokens": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "payment-info": {
            "$ref": "#/components/schemas/PaymentInfoDTO"
          }
        }
      },
      "ExtraInfoObjectDTO": {
        "type": "object",
        "properties": {
          "nav": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "tipoevento": {
            "type": "string"
          },
          "pa-emittente": {
            "type": "string"
          }
        }
      },
      "ExtraInfoResponseDTO": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtraInfoObjectDTO"
            }
          }
        }
      }
    }
  }
}